activity_sport_script/requirements.txt
pandas==2.2.2
numpy==1.26.4
faker==24.8.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
openpyxl==3.1.2
great_expectations==0.17.21